ContainerizedDaoTest class
